Change models to use short indices for ES 2.0 compliance
glDrawElements() on GL ES 2.0 does not support the GL_UNSIGNED_INT type
(support is enabled with extension GL_OES_element_index_uint).
This change modifies the OBJ loader and renderer to use 16-bit indices
so that the app works on contexts without that extension (specifically,
ANGLE on Android using the Vulkan backend).